+++ packages/moonstone/CHANGELOG.md
@@ -11,6 +11,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Fixed
 
 - `moonstone/Button` text color when used on a neutral (light) background in some cases
+- `moonstone/VideoPlayer` to no longer render its `children` in multiple places
 
 ## [1.0.0] - 2017-03-31