+++ packages/ui/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact ui module, newest changes on the top.
 
+## [unreleased]
+
+### Changed
+
+- `ui/Slottable` to support slot-candidate tags that have multiple props, which are now forwarded directly instead of just their children
+
 ## [1.0.0] - 2017-03-31
 
 ### Added